lexicon/requirements.txt
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==5.1.1
executing==1.2.0
idna==3.4
ipython==8.9.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
